// build/package/joint-plus.d.ts
    export namespace mvc {}

    export namespace attributes {}

    /* Extended JointJS namespaces */

    export namespace dia {

        namespace CommandManager {

            type ReduceOptionsCallback = (value: any, name: string) => boolean;

            interface Options {
                graph: dia.Graph;
                cmdBeforeAdd?: (eventName: string, ...eventArgs: any[]) => boolean;
                cmdNameRegex?: RegExp; /* a regular expression */
                applyOptionsList?: string[] | ReduceOptionsCallback;
                revertOptionsList?: string[] | ReduceOptionsCallback;
                storeReducedOptions?: boolean;
                stackLimit?: number;
                [key: string]: any;
            }

            interface EventOptions {
                [key: string]: any;
            }
            interface CommandData {
                id: string | number;
                type: string;
                previous: any;
                next: any;
                attributes: any;
            }

            interface Command {
                batch: boolean;
                action: string;
                data: CommandData;
                options: any;
                graphChange: boolean;
            }

            type BatchCommand = Command[];

            type Commands = Array<Command | BatchCommand>;

            interface JSON {
                undo: BatchCommand[];
                redo: BatchCommand[];
                [key: string]: any;
            }
        }

        class CommandManager extends mvc.Model<CommandManager.Options> {

            graph: dia.Graph;

            undoStack: CommandManager.Commands;

            redoStack: CommandManager.Commands;

            batchLevel: number | null;

            batchCommand: CommandManager.Commands | null;

            lastCmdIndex: number | null;

            undo(opt?: CommandManager.EventOptions): void;

            redo(opt?: CommandManager.EventOptions): void;

            cancel(opt?: CommandManager.EventOptions): void;

            reset(opt?: CommandManager.EventOptions): void;

            hasUndo(): boolean;

            hasRedo(): boolean;

            squashUndo(n?: number): void;

            squashRedo(n?: number): void;

            listen(): void;

            initBatchCommand(): void;

            storeBatchCommand(opt?: CommandManager.EventOptions): void;

            toJSON(): CommandManager.JSON;

            fromJSON(json: CommandManager.JSON, opt?: CommandManager.EventOptions): void;

            protected reduceOptions(opt: CommandManager.EventOptions): Partial<CommandManager.EventOptions>;

            protected exportBatchCommands(commands?: CommandManager.Commands): CommandManager.BatchCommand[];

            static sortBatchCommand(batchCommand: CommandManager.BatchCommand): CommandManager.BatchCommand;

            static filterBatchCommand(batchCommand: CommandManager.BatchCommand): CommandManager.BatchCommand;

            static squashCommands(commands: CommandManager.Commands, n?: number): CommandManager.Commands;
        }

        namespace Validator {

            export interface Options {
                commandManager: dia.CommandManager;
                cancelInvalid?: boolean;
            }

            type Callback = (err: Error, command: any, next: any) => any;
        }

        class Validator extends mvc.Model {

            constructor(opt: Validator.Options);

            options: Validator.Options;

            /* overrides mvc.Model.prototype.validate() */
            validate(actions: any, ...callback: Array<Validator.Callback>): Validator;
        }
    }
// ...
    export namespace ui {
// ...
        namespace Stencil {
// ...
            export interface Group {
                label: string | HTMLElement;
                index: number;
                closed?: boolean;
                height?: number;
                layout?: boolean | layout.GridLayout.Options | { [key: string]: any };
                paperOptions?: (() => dia.Paper.Options) | dia.Paper.Options;
                [key: string]: any;
            }
        }
// ...
    }